Implementing ProcTree logic
This commit brings a layer above the proc unit with the proc tree. A proc tree is an object that is responsible for processing proc units.
The goal is to be able to construct several proc trees and to changes them when we want to use a different pipeline (detecting a fence instead of a buoy).

This can be viewed as the filterchain of the proc_mapping (cf. proc_vision). There will be a service to change the current proc tree of the data interpreter.

To ease the modification of the proc tree, a configuration fil has been added containing the definition of all the proc tree and the proc unit parameters.