Implementing the ChangeProcTree callback.
The MapInterpreter is now a Subject<DetectionMode>, and send its detection mode to the observers.
As the SemanticMap is an observer, it will be able to process the correct algorithm for the newly detected objects (e.g. Banding box for the buoys).

The MapInterpreter provides a method for switching the DetectionMode (changing the proc tree).