First draft for the implementation of the map objects.
The cash system now takes MapObjects instead of keypoints.
The goal here is to abstract the algorithm for the detection/bounding box, weights, etc. an to put here in the specific map objects.
The code does not change for now, this commit is only about the structure. But a refactoring of the Semantic Map will be necessary as soon as we will get into the detection of other objects (wall, fence, etc.)