Completing the architecture to ease the detection of several object types.
We had to reconcider the way we handled the detection of the map object. The problem was that we were not able to store different object in the object registery (the wall is not a keypoint for instance).
To address this problem, a few architectural changes had to be implemented:

* The proc unit are know able to flow the data. Each output is redirected to the input of the next PU. This is done with a boost::any variant. Every PU has the responsability to cast the input to the correct type.
* The Odometry as well as the Coordinate system managment is now handled by a specific object called CoordinateSystems. It has been shared with the maps.
* The display of the semantic map is now done in the semantic map