Removing the Singleton pattern from the ObjectRegistery.
The object registery is now the real container for the semantic map object (and not a cache for the temporary object) as all the proccessing is now done in the PU (no proccessing is made in the semantic map anymore).
This commit allows the PU to manage the object in the semantic map more easily. They can now remove elements from the semantic map, or get the ROI of it.