First commit of 2.0 - Cleaning repo and base classes.
So since we don't do image proc in mapping now, I removed everything.
I created the sonarmapper and the base classes for it.

The semantic mapping logic will be reintegrate, but I prefer removing everything for the
freedom of deleting without breaking stuff.

Also, the sonar test currently test transformation from a point in meter to a local image point.